% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_elbo.R
\name{compute_elbo}
\alias{compute_elbo}
\title{Compute the objective function}
\usage{
compute_elbo(
  S = NULL,
  L = NULL,
  lambda = NULL,
  resid_s2,
  n,
  K = NULL,
  KL,
  R2 = NULL
)
}
\arguments{
\item{S}{An n-by-n Gram/covariance matrix}

\item{L}{An n-by-K matrix}

\item{lambda}{A K-length vector of weights}

\item{resid_s2}{The residual variance}

\item{n}{The number of samples}

\item{K}{The number of factors}

\item{KL}{A K-length vector containing the negative KL divergence}

\item{R2}{The expectation of the squared Frobenius norm of the residuals}
}
\value{
The objective function
}
\description{
Compute the objective function
}
