% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_R2.R
\name{compute_R2}
\alias{compute_R2}
\title{Compute the expectation of the squared Frobenius norm of the residuals}
\usage{
compute_R2(S, L, lambda, K)
}
\arguments{
\item{S}{An n-by-n Gram/covariance matrix}

\item{L}{An n-by-K matrix}

\item{lambda}{A K-length vector of weights}

\item{K}{The number of factors}
}
\value{
The expectation of the squared Frobenius norm of the residuals
}
\description{
Compute the expectation of the squared Frobenius norm of the residuals
}
