% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_resid_s2.R
\name{estimate_resid_s2}
\alias{estimate_resid_s2}
\title{Compute estimate for the residual variance}
\usage{
estimate_resid_s2(S = NULL, L = NULL, lambda = NULL, n, K = NULL, R2 = NULL)
}
\arguments{
\item{S}{An n-by-n Gram/covariance matrix}

\item{L}{An n-by-K matrix}

\item{lambda}{A K-length vector of weights}

\item{n}{The number of samples}

\item{K}{The number of factors}

\item{R2}{The expectation of the squared Frobenius norm of the residuals}
}
\value{
The residual variance
}
\description{
Compute estimate for the residual variance
}
