% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_means_loglik.R
\name{normal_means_loglik}
\alias{normal_means_loglik}
\title{Compute the normal means log likelihood}
\usage{
normal_means_loglik(x, s, Et, Et2)
}
\arguments{
\item{x}{A data vector}

\item{s}{A vector of standard errors}

\item{Et}{A vector of posterior means}

\item{Et2}{A vector of posterior second moments}
}
\value{
The log likelihood
}
\description{
Compute the normal means log likelihood
}
\references{
This function is copied from flashier:::normal.means.loglik
(https://github.com/willwerscheid/flashier/blob/master/R/objective.R)
}
