% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_factor.R
\name{optimize_factor}
\alias{optimize_factor}
\title{Optimize one factor}
\usage{
optimize_factor(
  R,
  ebnm_fn,
  maxiter,
  tol,
  v_init,
  lambda_k,
  g_k_init,
  R2k,
  n,
  KL
)
}
\arguments{
\item{R}{An n-by-n residual matrix}

\item{ebnm_fn}{An EBNM solver}

\item{maxiter}{The maximum number of iterations for the optimization}

\item{tol}{The convergence tolerance for the optimization}

\item{v_init}{An initial estimate for the factor}

\item{lambda_k}{An initial estimate for the factor weight, lambda}

\item{g_k_init}{An initial estimate for the prior; can be set to NULL}

\item{R2k}{The R2 value computed using the other factors}

\item{n}{The number of samples}

\item{KL}{The KL values for the other factors}
}
\value{
a proposed update for the factor estimate
}
\description{
Optimize one factor
}
