% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sym_ebcovmf_fit.R
\name{sym_ebcovmf_fit}
\alias{sym_ebcovmf_fit}
\title{Fit rank K symEBcovMF model}
\usage{
sym_ebcovmf_fit(S, ebnm_fn, Kmax, maxiter, rank_one_tol, tol)
}
\arguments{
\item{S}{A n-by-n Gram/covariance matrix}

\item{ebnm_fn}{An EBNM solver}

\item{Kmax}{The maximum number of factors to add}

\item{maxiter}{The maximum number of iterations used when optimizing the rank one fit}

\item{rank_one_tol}{The convergence tolerance for the rank one fit}

\item{tol}{The convergence tolerance for the rank K fit}
}
\value{
A symEBcovMF object
}
\description{
Fit rank K symEBcovMF model
}
