% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sym_ebcovmf_greedy.R
\name{sym_ebcovmf_greedy}
\alias{sym_ebcovmf_greedy}
\title{Fit a rank K symEBcovMF model by greedily adding factors}
\usage{
sym_ebcovmf_greedy(
  sym_ebcovmf_obj,
  ebnm_fn,
  Kmax,
  maxiter,
  rank_one_tol,
  tol,
  sign_constraint = "nonnegative",
  refit_lam = FALSE
)
}
\arguments{
\item{sym_ebcovmf_obj}{A symEBcovMF object}

\item{ebnm_fn}{An EBNM solver}

\item{Kmax}{The maximum number of factors to add}

\item{maxiter}{The maximum number of iterations used when optimizing the rank one fit}

\item{rank_one_tol}{The convergence tolerance for the rank one fit}

\item{tol}{The convergence tolerance for the rank K fit}

\item{sign_constraint}{The sign constraint (if any) for the loadings estimate}

\item{refit_lam}{True or False for if you want to refit the lambda values after each factor is added}
}
\value{
A symEBcovMF object
}
\description{
Fit a rank K symEBcovMF model by greedily adding factors
}
