% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sym_ebcovmf_init.R
\name{sym_ebcovmf_init}
\alias{sym_ebcovmf_init}
\title{Initialize rank K symEBcovMF Fit}
\usage{
sym_ebcovmf_init(S)
}
\arguments{
\item{S}{An n-by-n Gram/covariance matrix}
}
\value{
A symEBcovMF object
}
\description{
Initialize rank K symEBcovMF Fit
}
