% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nullcheck_factors.R
\name{nullcheck_factors}
\alias{nullcheck_factors}
\title{Check for factors that are zero vectors or have zero weight}
\usage{
nullcheck_factors(sym_ebcovmf_obj, L2_tol = 10^(-8))
}
\arguments{
\item{sym_ebcovmf_obj}{A symEBcovMF object}

\item{L2_tol}{The threshold for the L2 norm of the factor}
}
\value{
A symEBcovMF object
}
\description{
Check for factors that are zero vectors or have zero weight
}
