% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refit_lambda.R
\name{refit_lambda}
\alias{refit_lambda}
\title{Refit lambda}
\usage{
refit_lambda(S, sym_ebcovmf_obj, maxiter = 100, tol = 10^(-6))
}
\arguments{
\item{S}{An n-by-n Gram/covariance matrix}

\item{sym_ebcovmf_obj}{A symEBcovMF object}

\item{maxiter}{The maximum number of iterations for optimization}

\item{tol}{The convergence tolerance parameter for optimization}
}
\value{
A symEBcovMF object
}
\description{
Refit lambda
}
