% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sym_ebcovmf_backfit.R
\name{sym_ebcovmf_backfit}
\alias{sym_ebcovmf_backfit}
\title{Fit a rank K symEBcovMF model using backfit}
\usage{
sym_ebcovmf_backfit(
  sym_ebcovmf_obj,
  ebnm_fn,
  backfit_maxiter = 100,
  backfit_tol = 10^(-8),
  optim_maxiter = 500,
  optim_tol = 10^(-8)
)
}
\arguments{
\item{sym_ebcovmf_obj}{A symEBcovMF object}

\item{ebnm_fn}{An EBNM solver}

\item{backfit_maxiter}{The maximum number of iterations used when backfitting the rank K fit}

\item{backfit_tol}{The convergence tolerance used for the backfit}

\item{optim_maxiter}{The maximum number of iterations used when optimizing each factor}

\item{optim_tol}{The convergence tolerance used for optimizing each factor}
}
\value{
A symEBcovMF object
}
\description{
Fit a rank K symEBcovMF model using backfit
}
