% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sym_ebcovmf_factors_init.R
\name{sym_ebcovmf_factors_init}
\alias{sym_ebcovmf_factors_init}
\title{Initialize the factors at specified values}
\usage{
sym_ebcovmf_factors_init(S, sym_ebcovmf_obj, init_L, init_lambda)
}
\arguments{
\item{S}{A n-by-n Gram/covariance matrix}

\item{sym_ebcovmf_obj}{A symEBcovMF object}

\item{init_L}{An initial estimate for the loadings matrix. The columns of this matrix should have L2 norm equal to 1.}

\item{init_lambda}{A vector containing the initial values for the non-negative weights}
}
\value{
A symEBcovMF object
}
\description{
Initialize the factors at specified values
}
