% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sym_ebcovmf_r1_fit.R
\name{sym_ebcovmf_r1_fit}
\alias{sym_ebcovmf_r1_fit}
\title{Fit rank one symEBcovMF model}
\usage{
sym_ebcovmf_r1_fit(S, sym_ebcovmf_obj, ebnm_fn, maxiter, tol)
}
\arguments{
\item{S}{An n-by-n Gram/covariance matrix}

\item{sym_ebcovmf_obj}{A symEBcovMF object}

\item{ebnm_fn}{An EBNM solver}

\item{maxiter}{The maximum number of iterations used during optimization}

\item{tol}{The convergence tolerance parameter used during optimization}
}
\value{
A symEBcovMF object
}
\description{
Fit rank one symEBcovMF model
}
