% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sym_ebcovmf_r1_init.R
\name{sym_ebcovmf_r1_init}
\alias{sym_ebcovmf_r1_init}
\title{Initialize the greedily added factor}
\usage{
sym_ebcovmf_r1_init(R)
}
\arguments{
\item{R}{An n-by-n symmetric matrix}
}
\value{
An initial estimate for the factor
}
\description{
Initialize the greedily added factor
}
